@@ -17,6 +17,7 @@
     "@reduxjs/toolkit": "^2.3.0",
     "@types/yup": "^0.29.14",
     "axios": "^1.7.7",
+    "firebase": "^11.0.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-hook-form": "^7.53.1",
@@ -40,9 +41,9 @@
     "globals": "^15.9.0",
     "husky": "^8.0.0",
     "prettier": "^3.3.3",
-    "typescript": "^5.5.3",
+    "typescript": "^5.6.3",
     "typescript-eslint": "^8.7.0",
-    "vite": "^5.4.8",
+    "vite": "^5.4.10",
     "vite-plugin-eslint": "^1.8.1",
     "vite-plugin-svgr": "^4.2.0"
   }
