@@ -2,19 +2,19 @@ import React, { FC } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
 import { SubmitHandler, useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
-// import { useAppDispatch } from '../../utils/hooks/hooks';
+import { useAppDispatch } from '../../utils/hooks/hooks';
 import styles from './Auth.module.css';
 import { IFormData, IFormDataRegister } from '../../common/types/auth';
-
 import { LoginSchema, RegisterSchema } from '../../utils/yup/authForm';
 import SignIn from '../../components/AuthComponent/SignIn/SignIn';
 import SignUp from '../../components/AuthComponent/SignUp/SignUp';
 import Statistic from '../../components/AuthComponent/Statistic/Statistic';
+import { login, registerUser } from '../../redux/slices/auth/auth.thunks';
 
 const Auth: FC = () => {
   const location = useLocation();
   const navigate = useNavigate();
-  // const dispatch = useAppDispatch();
+  const dispatch = useAppDispatch();
   const {
     register,
     handleSubmit,
@@ -28,18 +28,19 @@ const Auth: FC = () => {
   const onFormSubmit: SubmitHandler<IFormData | IFormDataRegister> = async (
     data
   ) => {
-    console.log(data);
     if (location.pathname === '/login') {
       try {
-        // dispatch(login(data));
-        navigate('/');
+        const loginData = data as IFormData;
+        dispatch(login(loginData));
+        navigate('/dashboard');
       } catch (error) {
         console.error(error);
       }
     } else {
       try {
-        // dispatch(registerUser(data));
-        navigate('/');
+        const registerData = data as IFormDataRegister;
+        dispatch(registerUser(registerData));
+        navigate('/dashboard');
       } catch (error) {
         console.error(error);
       }
