@@ -3,14 +3,15 @@ import {
   selectIsLoading,
   selectIsLoggedIn,
   selectUser,
-  selectUserRole,
+  // selectUserRole,
 } from '../../redux/slices/auth/auth.selectors';
 
 export const useAuth = () => {
   const isLoggedIn = useAppSelector(selectIsLoggedIn);
   const isLoading = useAppSelector(selectIsLoading);
   const user = useAppSelector(selectUser);
-  const userRole = useAppSelector(selectUserRole);
+  const userRole = 'HR';
+  // useAppSelector(selectUserRole);
 
   return {
     isLoggedIn,
